@@ -1,66 +1,79 @@
 {
 	"name": "fable-frame",
-	"version": "0.7.2",
+	"version": "0.9.1",
 	"private": true,
 	"scripts": {
 		"dev": "vite dev",
-		"build": "vite build && cp -fr ./socketio/* build/",
-		"preview": "node build/server.js",
+		"build": "vite build",
+		"preview": "vite preview",
 		"zip": "npm run build; zip-build build .; mv *.zip builds/",
+		"sync": "svelte-kit sync",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"test:unit": "vitest",
 		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"format": "prettier --plugin-search-dir . --write .",
+		"build-apk": "android/gradlew assembleRelease",
+		"capacitor-sync": "npx cap sync"
 	},
 	"devDependencies": {
-		"@floating-ui/dom": "^1.4.5",
-		"@skeletonlabs/skeleton": "^2.1.0",
-		"@skeletonlabs/tw-plugin": "^0.2.0",
-		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/adapter-node": "^1.3.1",
-		"@sveltejs/kit": "^1.5.0",
-		"@tailwindcss/forms": "^0.5.4",
-		"@tailwindcss/typography": "^0.5.9",
-		"@types/node": "^20.6.0",
-		"@types/uuid": "^9.0.2",
-		"@typescript-eslint/eslint-plugin": "^5.45.0",
-		"@typescript-eslint/parser": "^5.45.0",
-		"autoprefixer": "^10.4.14",
-		"eslint": "^8.28.0",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-svelte": "^2.26.0",
-		"highlight.js": "^11.8.0",
-		"postcss": "^8.4.24",
-		"postcss-load-config": "^4.0.1",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.8.1",
-		"svelte": "^3.54.0",
-		"svelte-check": "^3.0.1",
-		"svelte-dnd-action": "^0.9.24",
+		"@achamaro/tailwindcss-iconify-icon": "^0.4.1",
+		"@floating-ui/dom": "^1.6.3",
+		"@skeletonlabs/skeleton": "^2.9.0",
+		"@skeletonlabs/tw-plugin": "^0.3.1",
+		"@sveltejs/adapter-auto": "^3.1.1",
+		"@sveltejs/adapter-node": "^5.0.1",
+		"@sveltejs/kit": "^2.5.4",
+		"@sveltejs/vite-plugin-svelte": "^3.0.2",
+		"@tailwindcss/forms": "^0.5.7",
+		"@tailwindcss/typography": "^0.5.10",
+		"@types/node": "^20.11.30",
+		"@types/uuid": "^9.0.8",
+		"@typescript-eslint/eslint-plugin": "^7.3.1",
+		"@typescript-eslint/parser": "^7.3.1",
+		"autoprefixer": "^10.4.19",
+		"eslint": "^8.57.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.35.1",
+		"highlight.js": "^11.9.0",
+		"postcss": "^8.4.38",
+		"postcss-load-config": "^5.0.3",
+		"prettier": "^3.2.5",
+		"prettier-plugin-svelte": "^3.2.2",
+		"svelte": "^4.2.12",
+		"svelte-check": "^3.6.8",
+		"svelte-dnd-action": "^0.9.42",
+		"svelte-icons": "^2.1.0",
 		"svelte-uicons": "^0.1.4",
-		"tailwindcss": "^3.3.3",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^4.3.0",
-		"vitest": "^0.25.3",
+		"tailwindcss": "^3.4.1",
+		"tslib": "^2.6.2",
+		"typescript": "^5.4.3",
+		"vite": "^5.2.2",
+		"vitest": "^1.4.0",
 		"zip-build": "^1.8.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"@dice-roller/rpg-dice-roller": "^5.4.1",
-		"@types/marked": "^5.0.1",
+		"@capacitor/android": "^5.7.4",
+		"@capacitor/cli": "^5.7.4",
+		"@capacitor/core": "^5.7.4",
+		"@dice-roller/rpg-dice-roller": "^5.5.0",
+		"@egoist/tailwindcss-icons": "^1.7.4",
+		"@iconify/json": "^2.2.194",
+		"@neodrag/svelte": "^2.0.3",
+		"@sveltejs/adapter-static": "^3.0.1",
+		"@types/marked": "^5.0.2",
 		"cors": "^2.8.5",
-		"dotenv": "^16.3.1",
-		"express": "^4.18.2",
-		"marked": "^7.0.5",
-		"pocketbase": "^0.15.2",
-		"pocketbase-typegen": "^1.1.11",
-		"socket.io": "^4.7.2",
-		"socket.io-client": "^4.7.2",
-		"svelte-icons": "^2.1.0",
+		"dotenv": "^16.4.5",
+		"express": "^4.19.1",
+		"js-base64": "^3.7.7",
+		"marked": "^12.0.1",
+		"pocketbase": "^0.21.1",
+		"pocketbase-typegen": "^1.2.1",
+		"svelte-disable-preload": "^0.0.2",
 		"svelte-icons-pack": "^2.1.0",
-		"svelte-media-queries": "^1.5.4",
-		"uuid": "^9.0.0"
+		"svelte-media-queries": "^1.6.2",
+		"svelte-resize-observer-action": "^0.0.3",
+		"uuid": "^9.0.1"
 	}
 }
