@@ -16,9 +16,9 @@
   
 </script>
 
-<BcSkillsTab on:fieldRemove on:fieldUpdate on:fieldAdd {compactVersion} {matches} fields={[...tabsContent["skill"]]} {activeTabName} {editMode} />
+<BcSkillsTab on:fieldRemove on:fieldUpdate on:fieldAdd {compactVersion} {matches} fields={[...tabsContent["skill"]]} tab={tabs["skill"]} {activeTabName} {editMode} />
 
-<BcFeelTab {character} on:fieldRemove on:fieldUpdate on:fieldAdd {compactVersion} {matches} fields={[...tabsContent["feel"]]} {activeTabName} {editMode} />
+<BcFeelTab {character} on:fieldRemove on:fieldUpdate on:fieldAdd {compactVersion} {matches} fields={[...tabsContent["feel"]]} {activeTabName} {editMode} tab={tabs["feel"]} />
 
 <!-- General -->
 <BcGeneralTab {character} on:fieldRemove on:fieldUpdate on:fieldAdd {compactVersion} tab={tabs["general"]} fields={[...tabsContent["general"]]} bind:activeTabName={activeTabName} {matches} {editMode} />
