@@ -1,6 +1,7 @@
 import type { Field } from "./types.d";
 import type { CharactersResponse } from "./pocketbase-types";
 import { v4 as uuidv4 } from 'uuid'
+import { getCharacter, updateCharacterWithHash } from "$models/character";
 
 export function getFieldFromListByName(name: string, fields: Field[]): Field {
   const returnedField = fields.find(field => field.name === name);
@@ -24,7 +25,7 @@ export function getFieldsByGroup(group: string, fields: Field[]): Field[] {
 }
 
 export function getCharacterFieldsByGroup(character: CharactersResponse, group: string): Field[] {
-  return character.fields.filter(field => field.group === group).sort((a, b) => {return a.weight - b.weight});
+  return character.fields.filter((field: Field) => field.group === group).sort((a: Field, b: Field) => {return a.weight - b.weight});
 }
 
 
@@ -36,22 +37,28 @@ export function addCharacterField(character: CharactersResponse, field: Field): 
   }
   field.weight = 1
 
-  character.fields?.push(field)
-  return character
+  const validateFieldById = character.fields?.find((f: Field) => f.id === field.id)
 
+  if (validateFieldById === undefined) {
+    character.fields?.push(field)
+    return character
+  }
+  else {
+    throw new Error(`Field with id ${field.id} already exists`)
+  }
 }
 
 export function removeCharacterField(character: CharactersResponse, field: Field): CharactersResponse {
 
   // Remove an item from character.fields Array comparing by id
-  character.fields = character.fields.filter(item => item.id!== field.id);
+  character.fields = character.fields.filter((f: Field) => f.id!== field.id);
 
   return character;
 }
 
 export function getCharacterFieldByName(character: CharactersResponse, fieldName: string): Field {
   
-  const field = character.fields?.find(f => f.name === fieldName);
+  const field = character.fields?.find((f: Field) => f.name === fieldName);
 
   if (field !== undefined) {
     return field;
@@ -69,7 +76,7 @@ export function getCharacterFieldByName(character: CharactersResponse, fieldName
 
 export function getCharacterFieldsByName(character: CharactersResponse, fieldName: string): Field[] {
   
-  const fields = character.fields.filter(f => f.name === fieldName);
+  const fields = character.fields.filter((f: Field) => f.name === fieldName);
 
   if (fields !== undefined) {
     return fields;
@@ -114,6 +121,36 @@ export function updateCharacterField(character: CharactersResponse, field: Field
     throw new Error(`Field ${field.id} not found in character`)
   }
 
+}
+
+export async function updateSaveCharacterField(characterId: string, field: Field): Promise<CharactersResponse> {
+
+  try {
+    const character = await getCharacter(characterId)
+    
+    updateCharacterField(character, field)
+    updateCharacterWithHash(characterId, character)
+    return character
+  }
+  catch (error) {
+    console.error(error)
+  }
+}
+
+export async function createCharacterField(characterId: string, field: Field, saveToCharacter: boolean = false): Promise<CharactersResponse> {
+
+  try {
+    let character = await getCharacter(characterId)
+    character = addCharacterField(character, field)
+
+    if (saveToCharacter) {
+      await updateCharacterWithHash(characterId, character)
+    }
+    return character
+  }
+  catch (error) {
+    console.error(error)
+  }
 }
 
 export function getFieldByNameFromList(list: Field[], name: string): Field {
