@@ -1,25 +1,87 @@
 <script lang="ts">
-	import type { CharactersResponse } from "$lib/pocketbase-types";
-	import { getCharacterAvatar } from "$models/character";
-	import { Avatar } from "@skeletonlabs/skeleton";
+	import type { CharactersResponse, RpgSystemsResponse } from "$lib/pocketbase-types";
+	import { toastShow } from "$lib/toast";
+	import { getBgCharacterImage } from "$models/character";
+	import { clipboard, getToastStore } from "@skeletonlabs/skeleton";
+	import { createEventDispatcher } from "svelte";
+	import { truncateText } from "$lib/utils";
+	import SquareCard from "../global/SquareCard.svelte";
+	import CircleIconButton from "../global/CircleIconButton.svelte";
+  import { Base64 } from 'js-base64';
+
+  const toastStore = getToastStore()
+  const dispatch = createEventDispatcher()
 
   export let character: CharactersResponse
+  export let rpgSystem: RpgSystemsResponse
   export let classes: string = ""
+  export let renderActionButtons: boolean = false
+
+  let bgCharacterImage = getBgCharacterImage(character, rpgSystem)
+
+	function getCharacterExport(): string {
+
+    const exportObject = {
+      rpgSystem: character.rpgSystem,
+      fields: character.fields,
+    }
+
+    // Base64 encode the JSON object
+    return Base64.encode(JSON.stringify(exportObject))
+	}
+
+  function cloneCharacterEvent(): void {
+    dispatch("cloneCharacter", {
+      characterId: character.id,
+    })
+  }
+
+  function deleteCharacterPromptEvent(): void {
+    dispatch("deleteCharacter", {
+      characterId: character.id,
+    })
+  }
   
 </script>
 
-<div class="flex flex-row {classes}">
-  <Avatar
-    initials={character.name.charAt(0)}
-    border="border-4 border-surface-300-600-token mr-3"
-    src={getCharacterAvatar(character)} />
-  <span class="flex flex-col justify-center">
-    <h3 class="h3 flex">{character.name}</h3>
-    {#if character.expand && character.expand.rpgSystem}
-      <article class="flex">{character.expand.rpgSystem.name}</article>
+<SquareCard
+  imageUrl={bgCharacterImage}
+  title={character.name}
+  link="/characters/{character.id}"
+  subtitle={rpgSystem !== undefined ? rpgSystem.name : ""}
+  {classes}
+  >
+  <svelte:fragment slot="actionButtons">
+    {#if renderActionButtons}
+      <CircleIconButton
+        on:click={cloneCharacterEvent}
+        icon="i-[la--clone-solid]"
+        color="variant-ghost-warning"
+      />
+      <button
+        use:clipboard={getCharacterExport()}
+        >
+        <CircleIconButton
+          on:click={() => toastShow("Character copied to clipboard", toastStore)}
+          icon="i-[mdi--export]"
+          color="variant-ghost-success"
+        />
+      </button>
+
+      <CircleIconButton
+        on:click={deleteCharacterPromptEvent}
+        icon="i-[material-symbols--delete]"
+        color="variant-ghost-error"
+      />
     {/if}
+
+  </svelte:fragment>
+
+  <svelte:fragment slot="linkButtons">
     {#if character.expand?.campaign !== undefined}
-      <span>Campaign: <a href="/campaigns/{character.expand.campaign.id}">{character.expand.campaign.name}</a></span>
+      <article class="mt-3">
+        <a class="btn variant-ghost-warning text-wrap" href="/campaigns/{character.expand.campaign.id}">{truncateText(character.expand.campaign.name, 25)}</a>
+      </article>
     {/if}
-  </span>
-</div>
+  </svelte:fragment>
+</SquareCard>
