@@ -1,5 +1,6 @@
 <script lang="ts">
 	import type { Field } from "$lib/types.d";
+	import { getModalStore, type ModalSettings } from "@skeletonlabs/skeleton";
 	import { createEventDispatcher } from "svelte";
 
   export let field: Field
@@ -8,6 +9,19 @@
   export let editMode: boolean = false
   export let labelStyle = ""
   export let valueStyle = ""
+  export let placeholder: string = ""
+  export let characterId: string = ""
+  export let color: string = ""
+  export let colorEdit: string = ""
+  export let colorButtons: string = ""
+  export let updateWithoutEditMode: boolean = false
+  export let placeholderArea: string = ""
+  export let editableClasses: string = ""
+  export let fullEditable: boolean = false
+  export let showTitle: boolean = true
+  export let isTextarea: boolean = false
+
+  const modalStore = getModalStore()
 
   let fieldValue: string = field.value
 
@@ -28,14 +42,36 @@
       field.value = fieldValueOld
     }
   }
+  
+  function openDescriptionModal() {
+
+    if (!editMode && field.description !== undefined) {
+
+      const descriptionModal: ModalSettings = {
+        type: 'alert',
+        title: "Description",
+        body: field.description,
+      }
+
+      modalStore.trigger(descriptionModal)
+    }
+  }
 
 </script>
 
-<label class="{classes} label">
-  <h4 class="h4 {labelStyle}">{field.label}</h4>
+<div class="{classes}">
+  
+  <button on:click={openDescriptionModal}
+    class="h3 mb-3 {labelStyle}">{field.label}
+  </button>
+
   {#if editable && editMode}
-    <input class="input" type="text" bind:value={fieldValue} on:focusout={fieldEdit} />
+    {#if isTextarea}
+      <textarea class="textarea" bind:value={fieldValue} on:focusout={fieldEdit} placeholder="{placeholder}" />
+    {:else}
+      <input class="input" type="text" bind:value={fieldValue} on:focusout={fieldEdit} placeholder="{placeholder}" />
+    {/if}
   {:else}
-    <p class="{valueStyle}">{field.value ?? ""}</p>
+    <p class="whitespace-pre-wrap {valueStyle}">{field.value ?? ""}</p>
   {/if}
-</label>
+</div>
