@@ -1,21 +1,25 @@
 <script lang="ts">
-	import type { CampaignsResponse, ScenesResponse } from "$lib/pocketbase-types";
+	import type { CampaignNotesResponse, CampaignsResponse, ScenesResponse } from "$lib/pocketbase-types";
 	import { getSceneImage } from "$models/scenes";
-	import { getDrawerStore, type DrawerSettings } from "@skeletonlabs/skeleton";
+	import { getDrawerStore, type DrawerSettings, ProgressBar } from "@skeletonlabs/skeleton";
 
-  // @ts-ignore
-	import GiPerson from 'svelte-icons/gi/GiPerson.svelte'
-  // @ts-ignore
-  import GiBookPile from 'svelte-icons/gi/GiBookPile.svelte'
-  // @ts-ignore
-  import MdChat from 'svelte-icons/md/MdChat.svelte'
 	import { currentUser } from "$lib/pocketbase";
+	import CampaignAlert from "../campaign/CampaignAlert.svelte";
+	import ScenesManagerCaller from "./ScenesManagerCaller.svelte";
+	import { getCampaignAlerts } from "$models/campaign_notes";
+	import { headerBanner, pageName } from "$lib/stores";
+	import ImageCanvasPan from "../global/ImageCanvasPan.svelte";
+	import { getCampaignImage } from "$models/campaign";
+	import Draggable from "../global/Draggable.svelte";
 
   export let scene: ScenesResponse
   export let campaign: CampaignsResponse
 
   let isGM = campaign.creator === $currentUser?.id || false
   let sceneImage = getSceneImage(scene)
+  let campaignAlerts: CampaignNotesResponse[] = []
+
+	pageName.set(`Scene: ${scene.name}`)
 
   const drawerStore = getDrawerStore();
 
@@ -31,9 +35,11 @@
   const characterDrawerSettings: DrawerSettings = {
     id: `campaign-character-list`,
     meta: {
-      campaignCharactersIds: isGM ? campaign.characters : campaign.expand.characters.filter(c => c.creator === $currentUser?.id).map(c => c.id)
+      rpgSystem: campaign.expand.rpgSystem,
+      campaign: campaign,
+      characters: isGM ? campaign.expand.characters : campaign.expand.characters.filter(c => c.creator === $currentUser?.id)
     },
-    width: "w-96",
+    width: "w-auto",
     position: "right",
   };
 
@@ -58,42 +64,77 @@
     drawerStore.open(campaignChatDrawerSettings);
   }
 
+  async function getCampaignAlertsHandler() {
+    campaignAlerts = await getCampaignAlerts(campaign.id)
+  }
+
+  $: {
+    if ($headerBanner !== getCampaignImage(campaign)) {
+      headerBanner.set(getCampaignImage(campaign))
+    }
+  }
+
 </script>
 
+<div class="bg-contain bg-center bg-no-repeat">
 
-<div class="h-full w-full bg-contain bg-center bg-no-repeat overflow-hidden"
-  style="background-image: url('{sceneImage}')">
+  <ImageCanvasPan
+    classes="absolute top-0 left-0 z-0"
+    imageUrl={sceneImage} height={window.innerHeight} />
 
+  <!-- Buttons -->
+  <div
+    class="fixed bottom-0 w-full justify-center flex">
+
+    <div
+      class="justify-center flex space-x-3 p-3
+        bg-surface-600/80 rounded-t-2xl shadow-md">
+      <!-- Campaign Notes -->
+      <button
+        class="btn btn-icon text-secondary-400"
+        on:click={openCampaignNotesDrawer}
+      >
+        <div class="i-[game-icons--book-pile] text-5xl"></div>
+      </button>
+      
+      <!-- Chat -->
+      <button
+        class="btn btn-icon text-secondary-400"
+        on:click={openCampaignChatDrawer}
+      >
+        <div class="i-[material-symbols--chat] text-5xl"></div>
+      </button>
+    
+      <!-- Scenes Manager -->
+      {#if isGM}
+        <ScenesManagerCaller
+          {campaign}
+          position="left"
+          classes="btn btn-icon text-secondary-400 text-5xl">
+        </ScenesManagerCaller>
+      {/if}
   
-  <div class="bg-surface-900/70 py-2">
-    <h2 class="h2 text-center flex justify-center space-x-3">
-      <p>{scene.name}</p>
-      <a class="btn btn-sm m-0 variant-ghost-secondary" href="/campaigns/{campaign.id}">{campaign.name}</a>
-    </h2>
+      <!-- Characters -->
+      <button
+        class="btn btn-icon text-secondary-400"
+        on:click={openCharactersDrawer}
+      >
+        <div class="i-[ri--profile-line] text-5xl"></div>
+      </button>
+    </div>
+
   </div>
 
-  <!-- Characters -->
-  <button
-    class="btn btn-icon variant-outline-primary fixed right-2 top-2/4 p-1"
-    on:click={openCharactersDrawer}
-  >
-    <GiPerson />
-  </button>
 
-  <!-- Campaign Notes -->
-  <button
-    class="btn btn-icon variant-outline-primary fixed left-2 top-2/4 p-1"
-    on:click={openCampaignNotesDrawer}
-  >
-    <GiBookPile />
-  </button>
-  
-  <!-- Character Notes (?) -->
-  <!-- Chat -->
-  <button
-    class="btn btn-icon variant-outline-primary fixed left-2 top-1/4 p-1"
-    on:click={openCampaignChatDrawer}
-  >
-    <MdChat />
-  </button>
+  <!-- Alerts -->
+  {#if isGM}
+    {#await getCampaignAlertsHandler()}
+      <ProgressBar />
+    {:then}
+      <Draggable title="Important">
+        <CampaignAlert bind:campaignAlerts={campaignAlerts} campaignId={campaign.id} />
+      </Draggable>
+    {/await}
+  {/if}
+
 </div>
