@@ -1,22 +1,22 @@
 <script lang="ts">
 	import { goto } from "$app/navigation";
+	import { getUploadedFileLocalSrc } from "$lib/getUploadedFileLocalSrc";
 	import type { ScenesResponse } from "$lib/pocketbase-types";
 	import { getScene, getSceneImage, createScene, updateScene } from "$models/scenes";
-  import { FileDropzone, ProgressBar } from "@skeletonlabs/skeleton";
+  import { FileDropzone, ProgressBar, getModalStore } from "@skeletonlabs/skeleton";
   export let campaignId: string
 
   export let sceneId: string = "new"
   
+  const modalStore = getModalStore()
+
   let scene: ScenesResponse
   let files: FileList
   let previewImageSrc: string
   let sceneEditImageSrc: string
 
   function fileSelectedHandler(e: any) {
-    // If no file was selected, empty the preview <img>
-    if(!e.target.files.length) return previewImageSrc = "";
-    // Set the <img>'s src to a reference URL to the selected file
-    return previewImageSrc = URL.createObjectURL(e.target.files.item(0))
+    previewImageSrc = getUploadedFileLocalSrc(e)
   }
 
   async function createSceneHandler() {
@@ -28,18 +28,20 @@
       formData.set("status", "0")
       formData.set("weight", "0")
 
-      if (files[0] !== undefined) {
+      if (files !== undefined && files[0] !== undefined) {
         formData.set("image", files[0])
+      }
+      try {
+        scene = await createScene(formData)
+        // @ts-ignore
+        $modalStore[0].response({scene: scene, action: "create"})
+        modalStore.close()
+      }
+      catch(error) {
+        console.log({error})
       }
     }
     
-    try {
-      scene = await createScene(formData)
-      goto(`/campaigns/${campaignId}/scenes`)
-    }
-    catch(error) {
-      console.log({error})
-    }
   }
 
   async function updateSceneHandler() {
@@ -49,13 +51,15 @@
       if (scene.name !== "") {
         formData.set("name", `${scene.name}`)
   
-        if (files[0] !== undefined) {
+        if (files !== undefined && files[0] !== undefined) {
           formData.set("image", files[0])
         }
       }
       try {
         scene = await updateScene(scene.id, formData)
-        goto(`/campaigns/${campaignId}/scenes`)
+        // @ts-ignore
+        $modalStore[0].response({scene: scene, action: "update"})
+        modalStore.close()
       }
       catch(error) {
         console.log({error})
@@ -73,6 +77,7 @@
       }
     }
     else {
+      // @ts-ignore
       scene = {
         id: sceneId,
         name: "",
@@ -80,6 +85,8 @@
         status: 0,
         weight: 0,
         campaign: campaignId,
+        group: "",
+        secret_hash: "",
       }
     }
   }
@@ -91,7 +98,7 @@
   <form class="flex flex-col items-center mx-auto space-y-3 w-96">
     <input
       class="input"
-      type="text" bind:value={scene.name} placeholder="Scene name"/>
+      type="text" bind:value={scene.name} placeholder="Scene name" required/>
 
     {#if sceneEditImageSrc !== "" || previewImageSrc !== ""}
       <img src={previewImageSrc !== undefined ? previewImageSrc : sceneEditImageSrc} alt="{scene.name}" />
