@@ -32,7 +32,6 @@
   // Reorder fields
   $: options = {
     items: fields,
-    flipDurationMs: flipDurationMs,
     dragDisabled: !editMode,
   }
 
@@ -47,7 +46,8 @@
       field.weight = counter
       counter++
       dispatch("fieldUpdate", {
-        field: field
+        field: field,
+        operation: "reorder",
       })
     })
     fields = e.detail.items
@@ -59,8 +59,7 @@
 </script>
 
 <CharacterSheetTabWrapper {compactVersion} {matches} {activeTabName}
-  {removable} {editMode}
-  tabName={tab.name}>
+  {removable} {editMode} {tab} on:fieldRemove >
   {#if editable && editMode}
   <input type="text" class="input h2 text-center" bind:value={tab.label} />
   {:else}
