@@ -3,28 +3,27 @@
 	import { page } from "$app/stores";
 	import Scene from "$lib/components/scenes/Scene.svelte";
   import type { CampaignsResponse, ScenesResponse } from "$lib/pocketbase-types";
-	import { getCampaignWithCharactersAndActiveScene, subscribeToCampaign } from "$models/campaign";
-	import { getScene } from "$models/scenes";
-	import { ProgressBar } from "@skeletonlabs/skeleton";
+	import { getCampaignWithCharactersAndActiveSceneAndRpgSystem, subscribeToCampaign } from "$models/campaign";
 	import type { UnsubscribeFunc } from "pocketbase";
 	import { onDestroy, onMount } from "svelte";
 
 
+  let campaignId: string = $page.params.campaignId
   let scene: ScenesResponse
   let campaign: CampaignsResponse
   let unsubscribeFromCampaign: UnsubscribeFunc
 
   async function loadData() {
     await setCampaignAndScene()
-    unsubscribeFromCampaign = await subscribeToCampaign($page.params.campaignId, testSub)
+    unsubscribeFromCampaign = await subscribeToCampaign(campaignId, campaignSub)
   }
 
   async function setCampaignAndScene() {
-    campaign = await getCampaignWithCharactersAndActiveScene($page.params.campaignId)
+    campaign = await getCampaignWithCharactersAndActiveSceneAndRpgSystem(campaignId)
     scene = campaign.expand.activeScene
   }
 
-  async function testSub({action, record}: {action: string, record: any}) {
+  async function campaignSub({action, record}: {action: string, record: any}) {
 
     if (action === "update") {
       if (campaign.activeScene !== record.activeScene) {
@@ -55,5 +54,8 @@
     <Scene {campaign} {scene} />
   {/key}
 {:else}
-  <ProgressBar />
+  <div class="flex flex-col items-center">
+      <h1 class="h1 text-center mb-3">No active scene. Come back later.</h1>
+      <a class="btn variant-filled-secondary" href="/campaigns/{campaignId}">Back to campaign</a>
+  </div>
 {/if}
