@@ -2,17 +2,17 @@
 	import { page } from "$app/stores";
 	import Scene from "$lib/components/scenes/Scene.svelte";
 import type { CampaignsResponse, ScenesResponse } from "$lib/pocketbase-types";
-	import { getCampaignWithCharacters } from "$models/campaign";
+	import { getCampaignWithCharactersAndActiveSceneAndRpgSystem } from "$models/campaign";
 	import { getScene } from "$models/scenes";
 	import { ProgressBar } from "@skeletonlabs/skeleton";
-
 
   let scene: ScenesResponse
   let campaign: CampaignsResponse
 
   async function loadData() {
     scene = await getScene($page.params.sceneId)
-    campaign = await getCampaignWithCharacters($page.params.campaignId)
+    campaign = await getCampaignWithCharactersAndActiveSceneAndRpgSystem($page.params.campaignId)
+
 }
 
 </script>
