@@ -1,7 +1,11 @@
 <script lang="ts">
+	import CreateCharacterChoiceCard from '$lib/components/characters/CreateCharacterChoiceCard.svelte';
   import barebones from "$lib/assets/barebones.webp"
   import wizard from "$lib/assets/wizard.webp"
+	import { pageName } from "$lib/stores";
+	import SquareCard from '$lib/components/global/SquareCard.svelte';
 
+	pageName.set("Create Character")
 </script>
 
 <div class="pt-3">
@@ -10,29 +14,25 @@
 
   <div class="flex justify-around flex-wrap">
 
-    <a href="/characters/create/wizard" class="card w-80 p-4 mb-3 justify-center items-center bg-initial card-hover">
-      <header>
-        <img src="{wizard}" alt="wizard" />
-      </header>
-      <div>
-        <h2 class="h2 text-center">Wizard</h2>
-        <article>
-          <p class="text-center">Create character step by step. Limited systems.</p>
-        </article>
-      </div>
-    </a>
+    <SquareCard
+      title="Wizard"
+      subtitle="Create character step by step. Limited systems."
+      imageUrl={wizard}
+      link="/characters/create/wizard"
+      />
 
-    <a href="/characters/create/barebones" class="card w-80 p-4 mb-3 justify-center items-center bg-initial card-hover">
-      <header>
-        <img src="{barebones}" alt="barebones" />
-      </header>
-      <div>
-        <h2 class="h2 text-center">Barebones</h2>
-        <article>
-          <p class="text-center">Build character sheet with custom fields from scratch</p>
-        </article>
-      </div>
-    </a>
+    <SquareCard
+      title="Barebones"
+      subtitle="Build character sheet with custom fields from scratch"
+      imageUrl={barebones}
+      link="/characters/create/barebones"
+      />
+
+    <SquareCard
+      title="Import"
+      subtitle="Import the previously exported character"
+      link="/characters/create/import"
+      />
 
   </div>
 </div>
